moved 'get_workdir_from_session' to context tools
