+++ openpype/hosts/fusion/utility_scripts/switch_ui.py
@@ -14,7 +14,7 @@ from openpype.pipeline import (
     legacy_io,
 )
 from openpype.hosts.fusion import api
-from openpype.lib.avalon_context import get_workdir_from_session
+from openpype.pipeline.context_tools import get_workdir_from_session
 
 log = logging.getLogger("Fusion Switch Shot")
